Complete converting write delay back to u8 revert

+++ src/lib.rs
@@ -224,7 +224,7 @@ mod register_access;
 pub struct Mlx9061x<I2C, IC> {
     /// The concrete I²C device implementation.
     i2c: I2C,
-    eeprom_write_delay_ms: u32,
+    eeprom_write_delay_ms: u8,
     address: u8,
     _ic: PhantomData<IC>,
 }
